Create Add Polygons to Points
Given an ESRI Shapefile and a .csv with lat/long values, pre-processes the data in order to add the polygon information to each lat/.long values.

Needs work: Assumes that the lat/lon values will be named "Latitude" and "Longitude". Does not allow user to choose which values of the polygon to add to the point file. Needs better instructions.